Fix typo (rename) in the name of ScheduledDateCalculator
