+++ src/entities/event/model/operations/index.ts
@@ -1,1 +1,1 @@
-export { default as ScheduledDateCalculator } from './SheduledDateCalculator';
+export { default as ScheduledDateCalculator } from './ScheduledDateCalculator';
